+++ _e2e/expect_test.ts
@@ -880,7 +880,7 @@ test("passes when mock object of last returned with", () => {
 test("passes when check object property via keyPath", () => {
   expect({ a: "b" }).toHaveProperty("a");
   expect({ a: { b: { c: "d" } } }).toHaveProperty("a.b.c");
-  expect({ a: { b: { c: "d" } } }).toHaveProperty(["a", "b", "c"]);
+  expect({ a: { b: { c: "d" } } }).not.toHaveProperty(["a", "b", "c"], "e");
 });
 
 test("passes when mock object returned successfully times", () => {